Unit tests + pdf converter added

+++ src/main/java/fr/diginamic/hello/services/DepartementService.java
@@ -30,8 +30,9 @@ public class DepartementService {
     private VilleRepository villeRepo;
 
     /**
-     * Demande au repository les départements contenus en base de données.
+     * Récupère les départements
      * @return liste de départements
+     * @throws RessourceNotFoundException aucun département à retourner
      */
     public List<Departement> getDepartements() throws RessourceNotFoundException {
         List<Departement> departements = departementRepo.findAll();
@@ -44,11 +45,13 @@ public class DepartementService {
     }
 
     /**
-     * Demande au repository les départements contenus en base de données, triés par nom.
-     * @param n nombre d'éléments à afficher
+     * Récupère les départements triés par nom et paginés.
+     * @param n nombre d'objets à afficher sur la page
      * @return liste de départements
+     * @throws RessourceNotFoundException aucun département à retourner
+     * @throws RequeteIncorrecteException les paramètres reçus en requête sont invalides
      */
-    public List<Departement> getDepartementsPagination(int n) throws RequeteIncorrecteException, RessourceNotFoundException {
+    public List<Departement> getDepartementsPagination(int n) throws RessourceNotFoundException, RequeteIncorrecteException {
         if (n <= 0) {
             throw new RequeteIncorrecteException("Le nombre d'éléments demandé doit être supérieur à 0.");
         }
@@ -64,11 +67,14 @@ public class DepartementService {
     }
 
     /**
-     * Méthode permettant de demander un département au repository à partir de son id.
+     * Récupère un département à partir de son ID.
+     * @param id identifiant du département
      * @return département
+     * @throws RessourceNotFoundException le département n'a pas pu être trouvé à partir de l'ID donné
+     * @throws RequeteIncorrecteException les paramètres reçus en requête sont invalides
      */
     public Departement getDepartementById(Long id) throws RessourceNotFoundException, RequeteIncorrecteException {
-        if (id == null) {
+        if (id == null || id < 0) {
             throw new RequeteIncorrecteException("Il faut renseigner un id.");
         }
 
@@ -83,8 +89,11 @@ public class DepartementService {
     }
 
     /**
-     * Demande un département au repository à partir de son code.
+     * Récupère un département à partir de son code.
+     * @param code code du département
      * @return département
+     * @throws RessourceNotFoundException le département n'a pas pu être trouvé à partir du code donné
+     * @throws RequeteIncorrecteException les paramètres reçus en requête sont invalides
      */
     public Departement getDepartementByCode(String code) throws RessourceNotFoundException, RequeteIncorrecteException {
         if (code == null || code.isEmpty()) {
@@ -102,8 +111,9 @@ public class DepartementService {
     }
 
     /**
-     * Méthode permettant de donner un département au repository à ajouter en base de données.
+     * Ajoute un département.
      * @param dept département à ajouter
+     * @throws RessourceExistanteException le département à ajouter existe déjà dans la base de données
      */
     public void insertDepartement(Departement dept) throws RessourceExistanteException {
         Optional<Departement> departementExistant = departementRepo.findByCode(dept.getCode());
@@ -116,14 +126,14 @@ public class DepartementService {
     }
 
     /**
-     * Méthode récupérant un département existant à partir de son id,
-     * et mettant-à-jour ses informations.
-     * Passage ensuite au repository pour persistance en base de données.
+     * Met-à-jour les données d'un département existant.
      * @param id identifiant du département à modifier
      * @param dept objet département contenant les nouvelles informations
+     * @throws RessourceNotFoundException le département n'a pas pu être trouvé à partir de l'ID donné
+     * @throws RequeteIncorrecteException les paramètres reçus en requête sont invalides
      */
     public void updateDepartement(Long id, Departement dept) throws RessourceNotFoundException, RequeteIncorrecteException {
-        if (id == null) {
+        if (id == null || id < 0) {
             throw new RequeteIncorrecteException("Il faut renseigner un id.");
         }
 
@@ -134,20 +144,21 @@ public class DepartementService {
         }
 
         Departement departementExistant = optDept.get();
-
         departementExistant.setNom(dept.getNom());
         departementExistant.setCode(dept.getCode());
         departementExistant.setVilles(dept.getVilles());
+
         departementRepo.save(departementExistant);
     }
 
     /**
-     * Méthode permettant de trouver le département existant à partir de son id,
-     * et de le donner au repository pour le supprimer.
+     * Supprimer un département et les villes de ce département.
      * @param id identifiant du département à supprimer
+     * @throws RessourceNotFoundException le département n'a pas pu être trouvé à partir de l'ID donné
+     * @throws RequeteIncorrecteException les paramètres reçus en requête sont invalides
      */
     public void deleteDepartement(Long id) throws RessourceNotFoundException, RequeteIncorrecteException {
-        if (id == null) {
+        if (id == null || id < 0) {
             throw new RequeteIncorrecteException("Il faut renseigner un id.");
         }
 
@@ -157,10 +168,14 @@ public class DepartementService {
             throw new RessourceNotFoundException(String.format("Aucun département dont l'Id est %d n'a été trouvé.", id));
         }
 
+        Departement departement = optDept.get();
+
         // Gère la relation one-to-many : si on supprime le département,
         // les villes liées sont supprimées aussi
-        for (Ville ville : optDept.get().getVilles()) {
-            villeRepo.deleteById(ville.getId());
+        if (departement.getVilles() != null) {
+            for (Ville ville : departement.getVilles()) {
+                villeRepo.deleteById(ville.getId());
+            }
         }
 
         departementRepo.deleteById(id);
